refactor(be): weight tinyint + document idempotency race in behavior events
weight fits tinyint (-3..5); note the find-then-write race in recordView/
recordReview is acceptable for fire-and-forget tracking.

+++ backend/src/behavior/behavior-event.entity.ts
@@ -28,7 +28,7 @@ export class UserProductEvent {
   })
   type!: BehaviorEventType;
 
-  @Column({ type: 'int' })
+  @Column({ type: 'tinyint' })
   weight!: number;
 
   @CreateDateColumn({ name: 'created_at', type: 'timestamp' })
